+++ src/middlewares/errorHandler.js
@@ -4,9 +4,7 @@ export const errorHandler = (err, req, res, next) => {
   } else {
     const status = err.status || (err.name === 'ZodError' ? 400 : 500);
     const message =
-      err?.issues?.[0]?.message ||
-      err.message ||
-      'Internal server error';
+      err?.issues?.[0]?.message || err.message || 'Internal server error';
 
     res.status(status).send({
       message,
